Adjust organization of compat layer
