+++ tests/socks4_no_auth.rs
@@ -1,6 +1,8 @@
 mod common;
 
 use common::*;
+#[cfg(feature = "futures-io")]
+use tokio_socks::io::Compat;
 use tokio_socks::{
     tcp::socks4::{Socks4Listener, Socks4Stream},
     Result,
@@ -49,9 +51,7 @@ fn bind_with_socket_no_auth() -> Result<()> {
 #[test]
 fn connect_with_socket_no_auth_futures_io() -> Result<()> {
     let runtime = futures_utils::runtime().lock().unwrap();
-    let socket = runtime
-        .block_on(futures_utils::connect_unix(UNIX_SOCKS4_PROXY_ADDR))?
-        .compat();
+    let socket = Compat::new(runtime.block_on(futures_utils::connect_unix(UNIX_SOCKS4_PROXY_ADDR))?);
     println!("socket connected");
     let conn = runtime.block_on(Socks4Stream::connect_with_socket(socket, ECHO_SERVER_ADDR))?;
     runtime.block_on(futures_utils::test_connect(conn))
@@ -62,9 +62,7 @@ fn connect_with_socket_no_auth_futures_io() -> Result<()> {
 fn bind_with_socket_no_auth_futures_io() -> Result<()> {
     let bind = {
         let runtime = futures_utils::runtime().lock().unwrap();
-        let socket = runtime
-            .block_on(futures_utils::connect_unix(UNIX_SOCKS4_PROXY_ADDR))?
-            .compat();
+        let socket = Compat::new(runtime.block_on(futures_utils::connect_unix(UNIX_SOCKS4_PROXY_ADDR))?);
         runtime.block_on(Socks4Listener::bind_with_socket(socket, ECHO_SERVER_ADDR))
     }?;
     futures_utils::test_bind_socks4(bind)
