+++ tests/socks4_userid.rs
@@ -1,6 +1,8 @@
 mod common;
 
 use common::*;
+#[cfg(feature = "futures-io")]
+use tokio_socks::io::Compat;
 use tokio_socks::{tcp::socks4::*, Result};
 
 #[cfg(feature = "tokio")]
@@ -61,9 +63,7 @@ fn bind_with_socket_userid() -> Result<()> {
 #[test]
 fn connect_with_socket_userid_futures_io() -> Result<()> {
     let runtime = futures_utils::runtime().lock().unwrap();
-    let socket = runtime
-        .block_on(futures_utils::connect_unix(UNIX_SOCKS4_PROXY_ADDR))?
-        .compat();
+    let socket = Compat::new(runtime.block_on(futures_utils::connect_unix(UNIX_SOCKS4_PROXY_ADDR))?);
     let conn = runtime.block_on(Socks4Stream::connect_with_userid_and_socket(
         socket,
         ECHO_SERVER_ADDR,
@@ -77,9 +77,7 @@ fn connect_with_socket_userid_futures_io() -> Result<()> {
 fn bind_with_socket_userid_futures_io() -> Result<()> {
     let bind = {
         let runtime = futures_utils::runtime().lock().unwrap();
-        let socket = runtime
-            .block_on(futures_utils::connect_unix(UNIX_SOCKS4_PROXY_ADDR))?
-            .compat();
+        let socket = Compat::new(runtime.block_on(futures_utils::connect_unix(UNIX_SOCKS4_PROXY_ADDR))?);
         runtime.block_on(Socks4Listener::bind_with_user_and_socket(
             socket,
             ECHO_SERVER_ADDR,
