+++ tests/username_auth.rs
@@ -1,6 +1,8 @@
 mod common;
 
 use common::*;
+#[cfg(feature = "futures-io")]
+use tokio_socks::io::Compat;
 use tokio_socks::{
     tcp::socks5::{Socks5Listener, Socks5Stream},
     Result,
@@ -68,7 +70,7 @@ fn bind_with_socket_username_auth() -> Result<()> {
 #[test]
 fn connect_with_socket_username_auth_futures_io() -> Result<()> {
     let runtime = futures_utils::runtime().lock().unwrap();
-    let socket = runtime.block_on(futures_utils::connect_unix(UNIX_PROXY_ADDR))?.compat();
+    let socket = Compat::new(runtime.block_on(futures_utils::connect_unix(UNIX_PROXY_ADDR))?);
     let conn = runtime.block_on(Socks5Stream::connect_with_password_and_socket(
         socket,
         ECHO_SERVER_ADDR,
@@ -83,7 +85,7 @@ fn connect_with_socket_username_auth_futures_io() -> Result<()> {
 fn bind_with_socket_username_auth_futures_io() -> Result<()> {
     let bind = {
         let runtime = futures_utils::runtime().lock().unwrap();
-        let socket = runtime.block_on(futures_utils::connect_unix(UNIX_PROXY_ADDR))?.compat();
+        let socket = Compat::new(runtime.block_on(futures_utils::connect_unix(UNIX_PROXY_ADDR))?);
         runtime.block_on(Socks5Listener::bind_with_password_and_socket(
             socket,
             ECHO_SERVER_ADDR,
